worklenz-backend/src-django/requirements.txt
asgiref==3.9.2
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
Markdown==3.9
pillow==12.0.0
reportlab==4.4.5
sqlparse==0.5.3
